requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
backoff==2.2.1
beautifulsoup4==4.13.3
blinker==1.9.0
cachetools==6.2.6
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
dataclasses-json==0.6.7
distro==1.9.0
duckduckgo_search==7.5.4
faiss-cpu==1.10.0
fastapi==0.115.11
fonttools==4.61.1
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.46
googleapis-common-protos==1.72.0
greenlet==3.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
importlib_metadata==8.7.1
Jinja2==3.1.6
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
langchain==1.2.10
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.11
langchain-openai==1.1.9
langchain-text-splitters==1.1.0
langfuse==3.9.1
langgraph==1.0.8
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.5
langsmith==0.7.1
lxml==6.0.2
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.1
multidict==6.7.1
mypy_extensions==1.1.0
narwhals==2.16.0
networkx==3.4.2
numpy==2.4.2
openai==2.7.2
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.7
ormsgpack==1.12.2
packaging==25.0
pandas==2.3.3
pillow==12.1.1
primp==0.15.0
propcache==0.4.1
protobuf==6.33.5
pyarrow==21.0.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydeck==0.9.1
pyparsing==3.3.2
pypdf==6.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2026.1.15
reportlab==4.4.10
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.30.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.46
sseclient==0.0.27
starlette==0.46.2
streamlit==1.51.0
tenacity==9.1.4
tiktoken==0.9.0
toml==0.10.2
tornado==6.5.4
tqdm==4.67.3
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uuid_utils==0.14.0
uvicorn==0.34.0
watchdog==6.0.0
wikipedia==1.4.0
wrapt==1.17.3
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
